Move NewsModelDataMapper to the presentation layer, where it belongs
